fix percent
