+++ packages/augur-ui/src/modules/market-cards/common.tsx
@@ -8,19 +8,21 @@ import {
   REPORTING_STATE,
   ZERO,
   INVALID_OUTCOME_ID,
+  YES_NO_YES_ID,
 } from 'modules/common/constants';
 import { createBigNumber, BigNumber } from 'utils/create-big-number';
 import ReactTooltip from 'react-tooltip';
 import TooltipStyles from 'modules/common/tooltip.styles.less';
 import { CheckCircleIcon } from 'modules/common/icons';
 import { OutcomeFormatted, FormattedNumber, MarketData } from 'modules/types';
-import { formatDai, formatAttoRep } from 'utils/format-number';
+import { formatDai, formatAttoRep, formatNumber } from 'utils/format-number';
 import { Getters } from '@augurproject/sdk';
 import { SecondaryButton } from 'modules/common/buttons';
 
 import Styles from 'modules/market-cards/common.styles.less';
 import MarketCard from 'modules/market-cards/market-card';
 import { selectSortedDisputingOutcomes } from 'modules/markets/selectors/market';
+import { calculatePosition } from 'modules/market/components/market-scalar-outcome-display/market-scalar-outcome-display';
 
 export interface PercentProps {
   percent: number;
@@ -166,21 +168,6 @@ export interface ScalarOutcomeProps {
   lastPrice?: FormattedNumber;
 }
 
-export function calculatePosition(
-  min: BigNumber,
-  max: BigNumber,
-  lastPrice: FormattedNumber
-) {
-  const range = max.minus(min);
-  const pricePercentage = createBigNumber(lastPrice ? lastPrice.value : 0)
-    .minus(min)
-    .dividedBy(range)
-    .times(createBigNumber(100))
-    .toNumber();
-
-  return lastPrice === null ? 50 : pricePercentage;
-}
-
 export const ScalarOutcome = (props: ScalarOutcomeProps) => (
   <div className={Styles.ScalarOutcome}>
     <div>
@@ -260,7 +247,7 @@ export const OutcomeGroup = (props: OutcomeGroupProps) => {
             min={props.min}
             max={props.max}
             lastPrice={
-              outcomesShow[0].price ? outcomesShow[0].lastPricePercent : null
+              outcomesShow[YES_NO_YES_ID].price ? formatNumber(outcomesShow[YES_NO_YES_ID].price) : null
             }
             scalarDenomination={props.scalarDenomination}
           />
